+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/commands/CommandSupplier.java
@@ -24,6 +24,7 @@ package com.github._1c_syntax.bsl.languageserver.commands;
 import org.eclipse.lsp4j.Command;
 
 import java.beans.Introspector;
+import java.util.List;
 import java.util.Optional;
 
 /**
@@ -63,8 +64,8 @@ public interface CommandSupplier<T extends CommandArguments> {
    * @param title Заголовок команды.
    * @return Команда с заполненными заголовком и идентификатором команды.
    */
-  default Command createCommand(String title) {
-    return new Command(title, getId());
+  default Command createCommand(String title, T arguments) {
+    return new Command(title, getId(), List.of(arguments));
   }
 
   /**
