+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/commands/ToggleCyclomaticComplexityInlayHintsCommandSupplier.java
@@ -21,6 +21,7 @@
  */
 package com.github._1c_syntax.bsl.languageserver.commands;
 
+import com.github._1c_syntax.bsl.languageserver.commands.complexity.AbstractToggleComplexityInlayHintsCommandSupplier;
 import com.github._1c_syntax.bsl.languageserver.inlayhints.CyclomaticComplexityInlayHintSupplier;
 import org.springframework.stereotype.Component;
